Add weights_only option to checkpoint loading
In some cases, loading checkpoints with safe_globals will not work.

Example: np.core was renamed to np._core in numpy>2.0.
Checkpoints that were saved with numpy<2.0 cannot be loaded in an environment with numpy>=2.0.
In that case, the only way to load the checkpoint is to set weights_only=False.